Add yarn.lock to image, and install dev deps for test
We need to install the yarn dev deps to run jest, but to do so,
we should use the yarn.lock. The Dockerfile is building a production
image, but we can include the yarn.lock, and use
`yarn install --production=false` to install the dev deps for running
jest with.