Build tester stage in Dockerfile
This will let us run Jest, as we'll be installing
all of our dev deps in this stage. We made sure to
dockerignore the node_modules directory as a minor perf
hack.

Also, now that we can run jest within this stage, we no
longer need to install dev deps in the `test` script in
package.json.